Add files via upload
Icosphere created so far. Horizontal and vertical connections have not been made thus far. The rendering made when the file is run gives colours corresponding to one of the angles from the spherical polar coordinates.  It seems that the main thing to do is to find the pentagons in the Icosphere. If we know what faces they are made from and what vertexes they are composed of we can define the horizontal and vertical connections. The pentagons are deviation points. To begin an impulse of excitation we need to start from one of these pentagons, for the sake of symmetry. If you could find the pentagons then it would be a help to me.