Poster things
Added new file and template, if you could add in the stuff and text you
think is pertinent then that is great. I couldn't get the smartdiagram
to work with the beamerposter package, so we will need to find another
way to do it. Also the attenuation of text is necessary